requirements.txt
django
python-decouple
python-telegram-bot
requests 
aiogram